Split ExperimentListView into its own file
